+++ src/main/java/tornadofx/Builders.kt
@@ -149,6 +149,22 @@ fun SplitPane.items(op: (Pane.() -> Unit)) {
     items.addAll(fake.children)
 }
 
+fun Pane.accordion(vararg panes: TitledPane, op: (Accordion.() -> Unit)? = null): Accordion {
+    var accordion = Accordion()
+    if (panes.isNotEmpty())
+        accordion.panes.addAll(panes)
+    opcr(this, accordion, op)
+    return accordion
+}
+
+fun Accordion.fold(title: String? = null, op: (Pane.() -> Unit)): TitledPane {
+    val vbox = VBox()
+    op(vbox)
+    val fold = TitledPane(title, if (vbox.children.size == 1) vbox.children[0] else vbox)
+    panes += fold
+    return fold
+}
+
 fun Pane.hbox(spacing: Double? = null, children: Iterable<Node>? = null, op: (HBox.() -> Unit)? = null): HBox {
     val hbox = HBox()
     if (children != null)
@@ -234,7 +250,7 @@ inline fun <S> Pane.treetableview(root: TreeItem<S>? = null, op: (TreeTableView<
     return treetableview
 }
 
-fun <S> TableView<S>.makeIndexColumn(name: String = "#", startNumber:Int = 1): TableColumn<S, Number> {
+fun <S> TableView<S>.makeIndexColumn(name: String = "#", startNumber: Int = 1): TableColumn<S, Number> {
     return TableColumn<S, Number>(name).apply {
         isSortable = false
         prefWidth = width